fixed logout null token issue

+++ client/src/features/goals/goalsSlice.js
@@ -14,8 +14,13 @@ const initialState = {
 export const createGoal = createAsyncThunk('goals/create',
     async (goalData, thunkAPI) => {
         try {
+
             const token = thunkAPI.getState().auth.user.token;
-            return goalService.createGoal(goalData, token)
+
+            if(token) {
+                return goalService.createGoal(goalData, token);
+            }
+            
         } catch (error) {
             const message = (error.response && error.response.data && error.response.data.message)
             || error.message || error.toString();
@@ -28,8 +33,16 @@ export const createGoal = createAsyncThunk('goals/create',
 export const getGoals = createAsyncThunk('goals/getAll',
     async (_, thunkAPI) => {
         try {
+
             const token = thunkAPI.getState().auth.user.token;
-            return goalService.getGoals(token)
+
+            if(token) {
+                return goalService.getGoals(token);
+            }
+            else {
+                console.log("user token is null")
+            }
+            
         } catch (error) {
             const message = (error.response && error.response.data && error.response.data.message)
             || error.message || error.toString();
@@ -42,8 +55,14 @@ export const getGoals = createAsyncThunk('goals/getAll',
 export const deleteGoal = createAsyncThunk('goals/delete',
     async (id, thunkAPI) => {
         try {
+
             const token = thunkAPI.getState().auth.user.token;
-            return await goalService.deleteGoal(id, token)
+
+            if(token) {
+                return await goalService.deleteGoal(id, token);
+            }
+
+    
         } catch (error) {
             const message = (error.response && error.response.data && error.response.data.message)
             || error.message || error.toString();
